filter_limit_values only needs 7 bits, make its tables smaller
Originally committed as revision 15499 to svn://svn.ffmpeg.org/ffmpeg/trunk

+++ libavcodec/vp3.c
@@ -230,7 +230,7 @@ typedef struct Vp3DecodeContext {
     int huff_code_size;
     uint16_t huffman_table[80][32][2];
 
-    uint32_t filter_limit_values[64];
+    uint8_t filter_limit_values[64];
     int bounding_values_array[256];
 } Vp3DecodeContext;
