+++ libavcodec/vp3data.h
@@ -84,7 +84,7 @@ static const uint32_t vp31_ac_scale_factor[64] =
    21,   19,   18,   17,   15,   13,  12,  10
 };
 
-static const uint32_t vp31_filter_limit_values[64] =
+static const uint8_t vp31_filter_limit_values[64] =
 {  30, 25, 20, 20, 15, 15, 14, 14,
    13, 13, 12, 12, 11, 11, 10, 10,
     9,  9,  8,  8,  7,  7,  7,  7,
